chore: refactor and test wip
